Non profit & academic friendly ratio of citations

Scientific journals operate over a broad spectrum of publishing strategies, from strictly for-profit, to non-profit, and in-between business models (e.g. for-profit but academic friendly journals).

From a list of references, this function computes three citation ratios: the proportion of non-profit citations, the proportion of for-profit and academic friendly citations, and the proportion of for-profit and non-academic friendly citations (Beck et al. 2025).

It uses the OpenAlex bibliographic database (https://openalex.org) to retrieve journal names from article DOI and the DAFNEE database (https://dafnee.isem-evolution.fr/) to get the business model and the academic friendly status of journals.

np_ratio(doi)

Arguments

doi

a character vector of Digital Object Identifiers (DOI). Can contain NA (book, book chapter, etc.).

Value

A list of two elements:

summary, a data.frame with two columns (metric and value) reporting the following statistics:

number of total references (length of doi argument)

number of references with DOI

number of deduplicated references

number of references found in the OpenAlex database

number of references whose journal is indexed in the DAFNEE database

number of non-profit and academic friendly references

number of for-profit and academic friendly references

number of for-profit and non academic friendly references

ratios, a vector of three ratios:

non-profit and academic friendly ratio

for-profit and academic friendly ratio

for-profit and non academic friendly ratio

References

Beck M et al. (2025) Strategic citations for a fairer academic landscape. Submitted to Proc B - Biological Science Practices.

Examples

# Import a BiBTex file (provided example) ----
filename <- system.file(
  file.path("extdata", "references.bib"),
  package = "faircitations"
)

refs <- RefManageR::ReadBib(filename)

length(refs)
#> [1] 38
refs[1:2]
#> [1] R. Johnson, A. Watkinson, and M. Mabe. “The STM report”. In: _An
#> overview of scientific and scholarly publishing. 5th edition October_
#> 94 (2018).
#> <https://policycommons.net/artifacts/1575771/2018_10_04_stm_report_2018/2265545/>.
#> 
#> [2] P. E. Smaldino and R. McElreath. “The natural selection of bad
#> science”. En. In: _Royal Society Open Science_ 3.9 (Sep. 2016), p.
#> 160384. ISSN: 2054-5703. DOI: 10.1098/rsos.160384.
#> <https://royalsocietypublishing.org/doi/10.1098/rsos.160384> (visited
#> on 07/10/2025).

# Extract DOI ----
doi_list <- unlist(refs$"doi")
names(doi_list) <- NULL
doi_list
#>  [1] "10.1098/rsos.160384"               "10.1126/science.1212540"          
#>  [3] "10.9745/GHSP-D-21-00145"           "10.1126/science.adk9900"          
#>  [5] "10.1016/j.ecolecon.2021.107082"    "10.1177/014107680609900316"       
#>  [7] "10.1093/reseval/rvad012"           "10.1111/ele.14395"                
#>  [9] "10.3998/ptpbio.3363"               "10.4000/proceedings.elpub.2018.30"
#> [11] "10.1093/scipol/scs093"             "10.1002/leap.1102"                
#> [13] "10.1087/0953151053584975"          "10.1177/0263395719858571"         
#> [15] "10.1017/S1062798709000532"         "10.1371/journal.pone.0127502"     
#> [17] "10.1162/qss_c_00305"               "10.1007/978-3-030-02511-3_1"      
#> [19] "10.1371/journal.pbio.1002264"      "10.3389/frma.2016.00007"          
#> [21] "10.48550/arXiv.2407.16551"         "10.1080/08109028.2014.891710"     
#> [23] "10.1007/s11192-022-04586-1"        "10.1098/rspb.2019.2047"           
#> [25] "10.5281/zenodo.4558704"            "10.1162/qss_a_00272"              
#> [27] "10.3917/inno.063.0095"             "10.1371/journal.pone.0243664"     
#> [29] "10.1257/jep.15.4.183"              "10.1073/pnas.0305628101"          
#> [31] "10.1016/j.amjmed.2019.07.028"      "10.32614/RJ-2023-089"             
#> [33] "10.5534/wjmh.230001"              

# Compute citation ratios ----
np_ratio(doi_list)
#>                                          metric value
#> 1                              N. of total refs  33.0
#> 2                             N. of refs w/ DOI  33.0
#> 3                      N. of refs w/o duplicate  33.0
#> 4                  N. of refs found in OpenAlex  30.0
#> 5                    N. of refs found in Dafnee  10.0
#> 6                         N. of Non-Profit refs   9.0
#> 7     N. of For-Profit refs (academic friendly)   1.0
#> 8 N. of For-Profit refs (non-academic friendly)   0.0
#> 9                              Non-Profit ratio   0.9